Changed delete events to use long instead of int
Streaming the public timeline tended to break very quickly before this fix

+++ Mastonet/StreamEventArgs.cs
@@ -17,6 +17,6 @@ namespace Mastonet
 
     public class StreamDeleteEventArgs : EventArgs
     {
-        public int StatusId { get; set; }
+        public long StatusId { get; set; }
     }
 }
